batman: Get `originator.c` to build

+++ sys/compat/linuxkpi/common/include/linux/jiffies.h
@@ -38,9 +38,9 @@
 #include <sys/kernel.h>
 #include <sys/limits.h>
 
-#define	jiffies			((unsigned long) ticks)
-#define	jiffies_64		((uint64_t) ticks)
-#define	jiffies_to_msecs(x)     (((int64_t)(int)(x)) * 1000 / hz)
+#define	jiffies			((unsigned long)ticks)
+#define	jiffies_64		((uint64_t)ticks)
+#define	jiffies_to_msecs(x)     ((unsigned int)(((int64_t)(int)(x)) * 1000 / hz))
 
 #define	MAX_JIFFY_OFFSET	((INT_MAX >> 1) - 1)
